--- RaidsPerRegion3/pom.xml
@@ -1,113 +1,0 @@
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>me.ShermansWorld</groupId>
-	<artifactId>RaidsPerRegion</artifactId>
-	<version>3.0.0</version>
-	<packaging>jar</packaging>
-	<name>RaidsPerRegion</name>
-	<url>http://maven.apache.org</url>
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.source>17</maven.compiler.source>
-		<maven.compiler.target>17</maven.compiler.target>
-	</properties>
-	<build>
-		<resources>
-			<resource>
-				<directory>resources</directory>
-			</resource>
-		</resources>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-shade-plugin</artifactId>
-				<version>3.4.1</version>
-				<configuration>
-					<createDependencyReducedPom>false</createDependencyReducedPom>
-					<minimizeJar>true</minimizeJar>
-					<relocations>
-						<relocation>
-							<pattern>fr.mrmicky.fastboard</pattern>
-							<shadedPattern>me.ShermansWorld.raidsperregion.fastboard</shadedPattern>
-						</relocation>
-					</relocations>
-				</configuration>
-				<executions>
-					<execution>
-						<goals>
-							<goal>shade</goal>
-						</goals>
-						<phase>package</phase>
-					</execution>
-				</executions>
-			</plugin>
-		</plugins>
-	</build>
-	<repositories>
-		<!-- This adds the Spigot Maven repository to the build -->
-		<repository>
-			<id>papermc</id>
-			<url>https://papermc.io/repo/repository/maven-public/</url>
-		</repository>
-		<repository>
-			<id>jitpack.io</id>
-			<url>https://jitpack.io</url>
-		</repository>
-		<repository>
-			<id>glaremasters repo</id>
-			<url>https://repo.glaremasters.me/repository/towny/</url>
-		</repository>
-		<repository>
-			<id>nexus</id>
-			<name>Lumine Releases</name>
-			<url>https://mvn.lumine.io/repository/maven-public/</url>
-		</repository>
-		<repository>
-			<id>enginehub-maven</id>
-			<url>https://maven.enginehub.org/repo/</url>
-		</repository>
-		<repository>
-			<id>sk89q-repo</id>
-			<url>https://maven.enginehub.org/repo/</url>
-		</repository>
-	</repositories>
-	<dependencies>
-		<!--This adds the Spigot API artifact to the build -->
-		<dependency>
-			<groupId>org.spigotmc</groupId>
-			<artifactId>spigot-api</artifactId>
-			<version>1.19.3-R0.1-SNAPSHOT</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.palmergames.bukkit.towny</groupId>
-			<artifactId>towny</artifactId>
-			<version>0.98.6.0</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
-			<groupId>io.lumine</groupId>
-			<artifactId>Mythic-Dist</artifactId>
-			<version>5.2.1</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.sk89q.worldedit</groupId>
-			<artifactId>worldedit-bukkit</artifactId>
-			<version>7.2.9</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.sk89q.worldguard</groupId>
-			<artifactId>worldguard-bukkit</artifactId>
-			<version>7.0.7</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
-			<groupId>fr.mrmicky</groupId>
-			<artifactId>fastboard</artifactId>
-			<version>1.2.1</version>
-		</dependency>
-	</dependencies>
-</project>
