+++ src/main/resources/config.yml
@@ -1,0 +1,69 @@
+# DO NOT CHANGE THIS
+config-version: 1
+
+# =========================================
+# |    RaidsPerRegion3 by ShermansWorld   |
+# =========================================
+
+RaidMobs:
+  SkeletalMinion:
+    Chance: 0.60
+    Priority: 1
+  StaticallyChargedSheep:
+    Chance: 0.35
+    Priority: 2
+  SkeletalKnight:
+    Chance: 0.15
+    Priority: 3
+  SkeletalKing:
+    Chance: 0.05
+    Priority: 4
+    
+Tier1:
+   MaxMobsAtOnce: 100
+   KillsGoal: 100
+   TimeLimit: 600
+   MobSpawnsPerCycle: 2
+   SpawnRate: 1.0
+   MobLevel: 1.0
+   HasBoss: false
+   Boss:
+Tier2:
+   MaxMobsAtOnce: 200
+   KillsGoal: 100
+   TimeLimit: 800
+   MobSpawnsPerCycle: 5
+   SpawnRate: 1.0
+   MobLevel: 1.0
+   HasBoss: false
+   Boss:
+Tier3:
+   MaxMobsAtOnce: 300
+   KillsGoal: 100
+   TimeLimit: 1000
+   MobSpawnsPerCycle: 8
+   SpawnRate: 1.0
+   MobLevel: 1.0
+   HasBoss: true
+   Boss: AngrySludge
+   
+
+# @OWNER, @TIER, @NAME, @BOSS
+RaidStartTitle: "&4&lTier @TIER Raid Inbound"
+RaidStartSubtitle: "&6Prepare to fight!" 
+
+RaidWinTitle: "&2&lRaid Won!"
+RaidWinSubtitle: "&6The raiders have fled"
+
+RaidLoseTitle: "&4&lRaid Lost!"
+RaidLoseSubtitle: "&6This can't be good..."
+
+RaidStopTitle: "&b&lRaid Cancelled"
+RaidStopSubtitle: "&6Raid Cancelled by @OWNER"
+
+RaidBossTitle: "&c&lRaid Boss Spawned"
+RaidBossSubtitle: "&6Kill @BOSS to win the raid!"
+
+PreventVanillaMobsSpawningInRaids: true
+ForceMobSpawningInRaids: true
+KeepMobsOnRaidLoss: true
